+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v0.3.0
 
 * Added threshold events
+* Support Y axis with units
+* Rework traces menu
 * Added possibility to control & readout position & size of plot window
 * Rework API with unified config as opposed to multiple `run_...` functions
 * Various UI improvements